Added test runner.
The test runner gives a graphical representation of the results of a bunch of tests.
Test go into "testrunner.js" into the "testCases" object. Simply add your test to the end of that object and it will automatically run your tests for you.

Please clone the class and the html file for each team if your team decides to use the test runner. or branch it off. If we need one single test runner for the entire group I can add in some subdivisions to the test runner and each group will have their own testrunner.js file.